requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.13.0
pydantic>=2.0.0
python-dotenv>=1.0.0